workers/cmd-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
transformers==4.36.2
torch==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
spacy==3.7.2
